New Updates :)

+++ application/views/x_layout.php
@@ -66,7 +66,10 @@ if(isset($_GET['go1'])){
 
         $view_sync_links = view_sync_links($i_fix['i__message'], true);
 
-        echo nl2br($i_fix['i__message']).'<br />Replace['.count($view_sync_links['replace_from']).']<br />';
+        if(count($view_sync_links['replace_from']) > 0){
+            echo nl2br($i_fix['i__message']).'<br />Replace['.count($view_sync_links['replace_from']).']<br />';
+        }
+
 
         foreach($view_sync_links['i__references'] as $ref_type => $ref_amount){
             if(count($ref_amount) > 0){
